+++ crates/transaction-pool/src/pool/best.rs
@@ -1,11 +1,11 @@
 use crate::{
     identifier::{SenderId, TransactionId},
     pool::pending::PendingTransaction,
-    PayloadTransactions, PoolTransaction, TransactionOrdering, ValidPoolTransaction,
+    PoolTransaction, TransactionOrdering, ValidPoolTransaction,
 };
-use alloy_consensus::Transaction;
 use alloy_primitives::Address;
 use core::fmt;
+use reth_payload_util::PayloadTransactions;
 use reth_primitives::TransactionSignedEcRecovered;
 use std::{
     collections::{BTreeMap, BTreeSet, HashSet, VecDeque},
@@ -20,7 +20,6 @@ use tracing::debug;
 /// This is a wrapper around [`BestTransactions`] that also enforces a specific basefee.
 ///
 /// This iterator guarantees that all transaction it returns satisfy both the base fee and blob fee!
-#[derive(Debug)]
 pub(crate) struct BestTransactionsWithFees<T: TransactionOrdering> {
     pub(crate) best: BestTransactions<T>,
     pub(crate) base_fee: u64,
@@ -73,7 +72,6 @@ impl<T: TransactionOrdering> Iterator for BestTransactionsWithFees<T> {
 /// be executed on the current state, but only yields transactions that are ready to be executed
 /// now. While it contains all gapless transactions of a sender, it _always_ only returns the
 /// transaction with the current on chain nonce.
-#[derive(Debug)]
 pub(crate) struct BestTransactions<T: TransactionOrdering> {
     /// Contains a copy of _all_ transactions of the pending pool at the point in time this
     /// iterator was created.
@@ -397,130 +395,6 @@ where
     }
 }
 
-/// An implementation of [`crate::traits::PayloadTransactions`] that yields
-/// a pre-defined set of transactions.
-///
-/// This is useful to put a sequencer-specified set of transactions into the block
-/// and compose it with the rest of the transactions.
-#[derive(Debug)]
-pub struct PayloadTransactionsFixed<T> {
-    transactions: Vec<T>,
-    index: usize,
-}
-
-impl<T> PayloadTransactionsFixed<T> {
-    /// Constructs a new [`PayloadTransactionsFixed`].
-    pub fn new(transactions: Vec<T>) -> Self {
-        Self { transactions, index: Default::default() }
-    }
-
-    /// Constructs a new [`PayloadTransactionsFixed`] with a single transaction.
-    pub fn single(transaction: T) -> Self {
-        Self { transactions: vec![transaction], index: Default::default() }
-    }
-}
-
-impl PayloadTransactions for PayloadTransactionsFixed<TransactionSignedEcRecovered> {
-    fn next(&mut self, _ctx: ()) -> Option<TransactionSignedEcRecovered> {
-        (self.index < self.transactions.len()).then(|| {
-            let tx = self.transactions[self.index].clone();
-            self.index += 1;
-            tx
-        })
-    }
-
-    fn mark_invalid(&mut self, _sender: Address, _nonce: u64) {}
-}
-
-/// Wrapper over [`crate::traits::PayloadTransactions`] that combines transactions from multiple
-/// `PayloadTransactions` iterators and keeps track of the gas for both of iterators.
-///
-/// We can't use [`Iterator::chain`], because:
-/// (a) we need to propagate the `mark_invalid` and `no_updates`
-/// (b) we need to keep track of the gas
-///
-/// Notes that [`PayloadTransactionsChain`] fully drains the first iterator
-/// before moving to the second one.
-///
-/// If the `before` iterator has transactions that are not fitting into the block,
-/// the after iterator will get propagated a `mark_invalid` call for each of them.
-#[derive(Debug)]
-pub struct PayloadTransactionsChain<B: PayloadTransactions, A: PayloadTransactions> {
-    /// Iterator that will be used first
-    before: B,
-    /// Allowed gas for the transactions from `before` iterator. If `None`, no gas limit is
-    /// enforced.
-    before_max_gas: Option<u64>,
-    /// Gas used by the transactions from `before` iterator
-    before_gas: u64,
-    /// Iterator that will be used after `before` iterator
-    after: A,
-    /// Allowed gas for the transactions from `after` iterator. If `None`, no gas limit is
-    /// enforced.
-    after_max_gas: Option<u64>,
-    /// Gas used by the transactions from `after` iterator
-    after_gas: u64,
-}
-
-impl<B: PayloadTransactions, A: PayloadTransactions> PayloadTransactionsChain<B, A> {
-    /// Constructs a new [`PayloadTransactionsChain`].
-    pub fn new(
-        before: B,
-        before_max_gas: Option<u64>,
-        after: A,
-        after_max_gas: Option<u64>,
-    ) -> Self {
-        Self {
-            before,
-            before_max_gas,
-            before_gas: Default::default(),
-            after,
-            after_max_gas,
-            after_gas: Default::default(),
-        }
-    }
-}
-
-impl<B, A> PayloadTransactions for PayloadTransactionsChain<B, A>
-where
-    B: PayloadTransactions,
-    A: PayloadTransactions,
-{
-    fn next(&mut self, ctx: ()) -> Option<TransactionSignedEcRecovered> {
-        while let Some(tx) = self.before.next(ctx) {
-            if let Some(before_max_gas) = self.before_max_gas {
-                if self.before_gas + tx.transaction.gas_limit() <= before_max_gas {
-                    self.before_gas += tx.transaction.gas_limit();
-                    return Some(tx);
-                }
-                self.before.mark_invalid(tx.signer(), tx.transaction.nonce());
-                self.after.mark_invalid(tx.signer(), tx.transaction.nonce());
-            } else {
-                return Some(tx);
-            }
-        }
-
-        while let Some(tx) = self.after.next(ctx) {
-            if let Some(after_max_gas) = self.after_max_gas {
-                if self.after_gas + tx.transaction.gas_limit() <= after_max_gas {
-                    self.after_gas += tx.transaction.gas_limit();
-                    return Some(tx);
-                }
-                self.after.mark_invalid(tx.signer(), tx.transaction.nonce());
-            } else {
-                return Some(tx);
-            }
-        }
-
-        None
-    }
-
-    fn mark_invalid(&mut self, sender: Address, nonce: u64) {
-        self.before.mark_invalid(sender, nonce);
-        self.after.mark_invalid(sender, nonce);
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -530,6 +404,7 @@ mod tests {
         Priority,
     };
     use alloy_primitives::U256;
+    use reth_payload_util::{PayloadTransactionsChain, PayloadTransactionsFixed};
 
     #[test]
     fn test_best_iter() {
