+++ crates/transaction-pool/src/traits.rs
@@ -886,7 +886,7 @@ impl<T> TransactionFilter for NoopTransactionFilter<T> {
     }
 }
 
-/// A Helper type that bundles the best transactions attributes together.
+/// A Helper type thaPayloadTransactionst bundles the best transactions attributes together.
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
 pub struct BestTransactionsAttributes {
     /// The base fee attribute for best transactions.
@@ -1508,24 +1508,6 @@ impl<Tx: PoolTransaction> Stream for NewSubpoolTransactionStream<Tx> {
             }
         }
     }
-}
-
-/// Iterator that returns transactions for the block building process in the order they should be
-/// included in the block.
-///
-/// Can include transactions from the pool and other sources (alternative pools,
-/// sequencer-originated transactions, etc.).
-pub trait PayloadTransactions {
-    /// Returns the next transaction to include in the block.
-    fn next(
-        &mut self,
-        // In the future, `ctx` can include access to state for block building purposes.
-        ctx: (),
-    ) -> Option<TransactionSignedEcRecovered>;
-
-    /// Exclude descendants of the transaction with given sender and nonce from the iterator,
-    /// because this transaction won't be included in the block.
-    fn mark_invalid(&mut self, sender: Address, nonce: u64);
 }
 
 #[cfg(test)]
